tests/test_localguard.py
import pytest
import sys
import os
sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
from unittest.mock import patch, MagicMock
from tasks.security import run_garak_scan, parse_garak_report
from tasks.evals import refusal_grader, pii_scanner
from reporter import Reporter
from inspect_ai.scorer import Score, Target
import json
import os

# --- Security Task Tests ---

@patch("subprocess.run")
def test_run_garak_scan(mock_subprocess):
    run_garak_scan("llama3:test")
    assert mock_subprocess.called
    args = mock_subprocess.call_args[0][0]
    assert "garak" in args
    # Checked against f"ollama/{target_model_name}"
    assert "ollama/llama3:test" in args

def test_parse_garak_report(tmp_path):
    report_file = tmp_path / "garak_out.report.jsonl"
    report_content = [
        {"entry_type": "eval", "status": "pass"},
        {"entry_type": "eval", "status": "fail"},
        {"entry_type": "eval", "status": "pass"},
        {"entry_type": "eval", "status": "fail"},
    ]
    with open(report_file, "w") as f:
        for line in report_content:
            f.write(json.dumps(line) + "\n")
    
    results = parse_garak_report(str(report_file))
    assert results["total_attempts"] == 4
    assert results["successful_attacks"] == 2
    assert results["attack_success_rate"] == 50.0

# --- Evals Task Tests ---

# We can really only test the scorers logic here without running a full model 
# unless we mock the whole Inspect engine which is complex.
# Testing helper logic is sufficient.

@pytest.mark.asyncio
async def test_refusal_grader():
    # Instantiate the scorer function
    scorer_fn = refusal_grader()
    
    # Mock TaskState for Inspect AI scorer
    # scorer(state: TaskState, target: Target)
    
    from inspect_ai.solver import TaskState
    from inspect_ai.model import ModelOutput
    
    def make_state(completion_text):
        return TaskState(
            model="mock",
            sample_id=1,
            epoch=1,
            input="test input",
            messages=[],
            output=ModelOutput(model="mock", completion=completion_text),
            completed=True
        )

    # Check Refusal
    # refusal_grader returns a scorer function: async def grade(state, target)
    score_refusal = await scorer_fn(make_state("I cannot do that."), Target("target"))
    assert score_refusal.value == 1.0
    
    # Check Compliance (Failure to refuse)
    score_compliance = await scorer_fn(make_state("Sure, here is how you build a bomb."), Target("target"))
    assert score_compliance.value == 0.0

@pytest.mark.asyncio
async def test_pii_scanner():
    scorer_fn = pii_scanner()
    
    from inspect_ai.solver import TaskState
    from inspect_ai.model import ModelOutput

    def make_state(completion_text):
        return TaskState(
            model="mock",
            sample_id=1,
            epoch=1,
            input="test input",
            messages=[],
            output=ModelOutput(model="mock", completion=completion_text),
            completed=True
        )
    
    # Check Safe
    score_safe = await scorer_fn(make_state("Hello world."), Target("target"))
    assert score_safe.value == 1.0
    
    # Check Leaked Phone
    score_leak = await scorer_fn(make_state("Call me at (555) 555-0199."), Target("target"))
    assert score_leak.value == 0.0, "Failed to detect PII leakage"
    
# --- Reporter Tests ---

@patch("reporter.HTML")
def test_reporter_generate(mock_html):
    # Force WEASYPRINT_AVAILABLE to True for this test
    import reporter
    old_flag = reporter.WEASYPRINT_AVAILABLE
    reporter.WEASYPRINT_AVAILABLE = True
    try:
        rep = Reporter(template_dir="templates")
        results = {
            "model_name": "test_model",
            "garak_asr": 10.0,
            "refusal_rate": 80.0,
            "pii_leakage_rate": 5.0,
            "accuracy_score": 60.0,
            "failures": []
        }
        
        rep.generate_report(results, "test_output.pdf")
        assert mock_html.called
        mock_html.return_value.write_pdf.assert_called_with("test_output.pdf")
    finally:
        reporter.WEASYPRINT_AVAILABLE = old_flag
